mcp/pyproject.toml
[project]
name = "mineru-open-mcp"
version = "1.0.18"
description = "MinerU MCP Server for PDF to Markdown conversion"
authors = [
    {name = "minerU",email = "OpenDataLab@pjlab.org.cn"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastmcp>=3.1.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "uvicorn>=0.20.0",
    "starlette>=0.27.0",
    "mineru-open-sdk>=0.1.1",
]

[project.scripts]
mineru-open-mcp = "mineru_open_mcp.cli:main"

[tool.poetry]
packages = [{include = "mineru_open_mcp", from = "src"}]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

